Move formatter settings form to base class
